Run ermineJ analysis

The monolith function that wraps the emrineJ cli. It is recommended to use the wrapper functions (ora, gsr, corr, roc).

ermineR(annotation = NULL, aspects = c("Molecular Function",
  "Cellular Component", "Biological Process"), scores = NULL,
  hitlist = NULL, scoreColumn = 1, threshold = 0.001, expression = NULL,
  bigIsBetter = FALSE, customGeneSets = NULL, geneReplicates = c("mean",
  "best"), logTrans = FALSE, pAdjust = c("FDR", "FWE"), test = c("ORA",
  "GSR", "CORR", "ROC"), iterations = NULL, stats = c("mean", "quantile",
  "meanAboveQuantile", "precisionRecall"), quantile = 50,
  geneSetDescription = "Latest_GO", output = NULL, return = TRUE,
  minClassSize = 20, maxClassSize = 200)

Arguments

annotation
Annotation. A file path, a data.frame or a platform short name (eg. GPL127). If given a platform short name it will be downloaded from annotation repository of Pavlidis Lab (www.chibi.ubc.ca/microannots/). Note that if there is a file or folder with the same name as the platform name in the directory, that file will be read instead of getting a copy from Pavlidis Lab. If this file isn't a valid annotation file, the function will fail. If providing a custom annotation file, see makeAnnotation to do it from R or erminej.msl.ubc.ca/help/input-files/gene-annotations/ to do it manually.

If you are providing a custom gene set, you can leave annotation as NULL

aspects
Character vector. Which Go aspects to include in the analysis. Can be in long form (eg. 'Molecular Function') or short form (eg. c('M','C','B'))

scores
A data.frame. Rownames have to be gene identifiers (eg. probes, must be unique), followed by any number of columns. The column used for scoring is chosen by scoreColumn. See http://erminej.msl.ubc.ca/help/input-files/gene-scores/ for information abot how to specify scores. (for test = ORA, GSR and ROC)

scoreColumn
Integer or character. Which column of the scores data.frame to use as scores. Defaults to first column of scores. See http://erminej.msl.ubc.ca/help/input-files/gene-scores/ for details. (for test = ORA, GSR and ROC)

threshold
Double. Score threshold (test = ORA only)

expression
A file path or a data frame. Expression data. (test = CORR only) Necesary correlation anaylsis. See http://erminej.msl.ubc.ca/help/input-files/gene-expression-profiles/ for data format

bigIsBetter
Logical. If TRUE large scores are considered to be higher. FALSE by default (as in p values).

customGeneSets
Path to a directory that contains custom gene set files, paths to custom gene set files themselves or a named list of character strings. Use this option to create your own gene sets. If you provide directory you can specify probes or gene symbols to include in your gene sets. See http://erminej.msl.ubc.ca/help/input-files/gene-sets/ for information about format for this file. If you are providing a list, only gene symbols are accepted.

geneReplicates
What to do when genes have multiple scores in input file (due to multiple probes per gene)

logTrans
Logical. Should the data be -log10 transformed. Recommended for p values. FALSE by default

pAdjust
Which multiple test correction method to use. Can be "FDR" or 'Westfall-Young' (slower).

test
Method for computing gene set significance

iterations
Number of iterations. We suggest a starting value of 10000 iterations. When you decide on parameters you like, we recommend a larger number of iterations (perhaps 200,000 or more). This is to get sufficient precision in the p-values to make multiple-test correction work correctly. (test = GSR CORR and precRecall methods only)

stats
Method for computing raw class statistics (test = GSR only)

quantile
Integer. Quantile to use. (stats = meanAboveQuantile only)

geneSetDescription
"Latest_GO", a file path that leads to a GO XML or OBO file or a URL that leads to a go ontology file that ends with rdf-xml.gz.

If you left annotation as NULL and provided customGeneSets, this argument is not required and will default to NULL. Otherwise, by default it'll be set to "Latest_GO" which downloads the latest available GO XML file. This option won't work without an internet connection. To get a frozen file that you can use later, see goToday, goAtDate and getGoDates. See http://erminej.msl.ubc.ca/help/input-files/gene-set-descriptions/ for details.

output
Output file name.

return
If results should be returned. Set to FALSE if you only want a file

minClassSize
minimum class size

maxClassSize
maximum class size

Value

A list