GSR

The goal of this method is the same as for ora: to provide a p value for each gene set. The key difference lies in that ORA requires that you select a threshold for “gene selection”, whereas GSR does not.

GSR uses all the gene scores for the genes in a gene set to produce a score. This means that genes that do not meet a statistical threshold for selection can contribute to the score. In addition, more information contained in the gene scores is preserved than in ORA, because ORA is essentially rank-based, whereas GSR uses the gene scores themselves. (The precision-recall method is even more close to the ORA method).

In practice, ORA and GSR can yield similar results; however, we have found that GSR tends to be more robust than ORA (because there is no threshold to set) and can give interesting results in situations where ORA doesn’t work as well (when no genes meet the predetermined selection threshold).

GSR is appropriate when you have reasonably good trust in the values of your gene scores, as opposed to the ranking. Its key advantage over ORA is that you do not have to set a threshold gene score. If you trust the ranking of gene scores more, the precision-recall or ROC methods might be useful.

Method overview taken from: http://erminej.msl.ubc.ca/help/tutorials/running-an-analysis-resampling/

gsr(scores, scoreColumn = 1, bigIsBetter = FALSE, logTrans = FALSE,
  annotation = NULL, aspects = c("Molecular Function", "Cellular Component",
  "Biological Process"), iterations, geneReplicates = c("mean", "best"),
  pAdjust = c("FDR", "Bonferroni"), geneSetDescription = "Latest_GO",
  customGeneSets = NULL, minClassSize = 20, maxClassSize = 200,
  output = NULL, return = TRUE)

Arguments

scores
A data.frame. Rownames have to be gene identifiers (eg. probes, must be unique), followed by any number of columns. The column used for scoring is chosen by scoreColumn. See http://erminej.msl.ubc.ca/help/input-files/gene-scores/ for information abot how to specify scores. (for test = ORA, GSR and ROC)

scoreColumn
Integer or character. Which column of the scores data.frame to use as scores. Defaults to first column of scores. See http://erminej.msl.ubc.ca/help/input-files/gene-scores/ for details. (for test = ORA, GSR and ROC)

bigIsBetter
Logical. If TRUE large scores are considered to be higher. FALSE by default (as in p values).

logTrans
Logical. Should the data be -log10 transformed. Recommended for p values. FALSE by default

annotation
Annotation. A file path, a data.frame or a platform short name (eg. GPL127). If given a platform short name it will be downloaded from annotation repository of Pavlidis Lab (www.chibi.ubc.ca/microannots/). Note that if there is a file or folder with the same name as the platform name in the directory, that file will be read instead of getting a copy from Pavlidis Lab. If this file isn't a valid annotation file, the function will fail. If providing a custom annotation file, see makeAnnotation to do it from R or erminej.msl.ubc.ca/help/input-files/gene-annotations/ to do it manually.

If you are providing a custom gene set, you can leave annotation as NULL

aspects
Character vector. Which Go aspects to include in the analysis. Can be in long form (eg. 'Molecular Function') or short form (eg. c('M','C','B'))

iterations
Number of iterations (test = GSR and CORR methods only)

geneReplicates
What to do when genes have multiple scores in input file (due to multiple probes per gene)

pAdjust
Which multiple test correction method to use. Can be "FDR" or 'Westfall-Young' (slower).

geneSetDescription
"Latest_GO", a file path that leads to a GO XML or OBO file or a URL that leads to a go ontology file that ends with rdf-xml.gz.

If you left annotation as NULL and provided customGeneSets, this argument is not required and will default to NULL. Otherwise, by default it'll be set to "Latest_GO" which downloads the latest available GO XML file. This option won't work without an internet connection. To get a frozen file that you can use later, see goToday, goAtDate and getGoDates. See http://erminej.msl.ubc.ca/help/input-files/gene-set-descriptions/ for details.

customGeneSets
Path to a directory that contains custom gene set files, paths to custom gene set files themselves or a named list of character strings. Use this option to create your own gene sets. If you provide directory you can specify probes or gene symbols to include in your gene sets. See http://erminej.msl.ubc.ca/help/input-files/gene-sets/ for information about format for this file. If you are providing a list, only gene symbols are accepted.

minClassSize
minimum class size

maxClassSize
maximum class size

output
Output file name.

return
If results should be returned. Set to FALSE if you only want a file