Precision-recall

Precision-recall curves are a standard way to analyze the quality of a classifier, as are roc curves. The difference between ROC and Precision-recall (PR) is that PR is mostly concerned with what is happening at the top of the ranking, while ROC looks at the whole ranking. The precision-recall method in ErmineJ is similar in intent to the method popularized in “GSEA“, but uses standard precision-recall (GSEA originally used Kolmogorov-Smirnov statistics, but was changed to use a modified K-S statistic that makes it more precision-recall-like).

Like the ROC method, the PRC method uses only the ranks of the gene scores. That is, all it cares about is the ordering of items obtained by your gene scores (e.g., t-test or fold-change), but doesn’t use the information about the relative values of the scores.

PR-scoring is rather like the ORA method in that it is concerned with what is going on at the “top” of your list, but unlike ORA it does not use a threhold. Thus PR-scoring would be appropriate whenever you consider the ORA method appropriate. The PR method requires a resampling step, so it is slower to run than ORA.

Method overview taken from: http://erminej.msl.ubc.ca/help/tutorials/tutorial-precision-recall-scoring/

precRecall(scores, scoreColumn = 1, bigIsBetter = FALSE, logTrans = FALSE,
  annotation = NULL, aspects = c("Molecular Function", "Cellular Component",
  "Biological Process"), iterations, geneReplicates = c("mean", "best"),
  pAdjust = c("FDR", "Bonferroni"), geneSetDescription = "Latest_GO",
  customGeneSets = NULL, minClassSize = 20, maxClassSize = 200,
  output = NULL, return = TRUE)

Arguments

scores
A data.frame. Rownames have to be gene identifiers (eg. probes, must be unique), followed by any number of columns. The column used for scoring is chosen by scoreColumn. See http://erminej.msl.ubc.ca/help/input-files/gene-scores/ for information abot how to specify scores. (for test = ORA, GSR and ROC)

scoreColumn
Integer or character. Which column of the scores data.frame to use as scores. Defaults to first column of scores. See http://erminej.msl.ubc.ca/help/input-files/gene-scores/ for details. (for test = ORA, GSR and ROC)

bigIsBetter
Logical. If TRUE large scores are considered to be higher. FALSE by default (as in p values).

logTrans
Logical. Should the data be -log10 transformed. Recommended for p values. FALSE by default

annotation
Annotation. A file path, a data.frame or a platform short name (eg. GPL127). If given a platform short name it will be downloaded from annotation repository of Pavlidis Lab (www.chibi.ubc.ca/microannots/). Note that if there is a file or folder with the same name as the platform name in the directory, that file will be read instead of getting a copy from Pavlidis Lab. If this file isn't a valid annotation file, the function will fail. If providing a custom annotation file, see makeAnnotation to do it from R or erminej.msl.ubc.ca/help/input-files/gene-annotations/ to do it manually.

If you are providing a custom gene set, you can leave annotation as NULL

aspects
Character vector. Which Go aspects to include in the analysis. Can be in long form (eg. 'Molecular Function') or short form (eg. c('M','C','B'))

iterations
Number of iterations (test = GSR and CORR methods only)

geneReplicates
What to do when genes have multiple scores in input file (due to multiple probes per gene)

pAdjust
Which multiple test correction method to use. Can be "FDR" or 'Westfall-Young' (slower).

geneSetDescription
"Latest_GO", a file path that leads to a GO XML or OBO file or a URL that leads to a go ontology file that ends with rdf-xml.gz. Note that this is a mandatory argument. It defaults to Latest_GO but that option won't work if you don't have a working internet connection. To get a frozen file that you can use later, see goToday, goAtDate and getGoDates. See http://erminej.msl.ubc.ca/help/input-files/gene-set-descriptions/ for details.

Can be set to NULL if annotation is also NULL and a cutsom gene set is provided.

customGeneSets
Path to a directory that contains custom gene set files, paths to custom gene set files themselves or a named list of character strings. Use this option to create your own gene sets. If you provide directory you can specify probes or gene symbols to include in your gene sets. See http://erminej.msl.ubc.ca/help/input-files/gene-sets/ for information about format for this file. If you are providing a list, only gene symbols are accepted.

minClassSize
minimum class size

maxClassSize
maximum class size

output
Output file name.

return
If results should be returned. Set to FALSE if you only want a file