ROC

The receiver operator characteristic (ROC) method is a fast, non-parametric alternative to the ORA and resampling methods for generating gene set scores from gene scores.

roc(scores, scoreColumn = 1, bigIsBetter = FALSE, logTrans = FALSE,
  annotation, aspects = c("Molecular Function", "Cellular Component",
  "Biological Process"), geneReplicates = c("mean", "best"),
  pAdjust = c("FDR", "Bonferroni"), geneSetDescription = "Latest_GO",
  customGeneSets = NULL, minClassSize = 20, maxClassSize = 200,
  output = NULL, return = TRUE)

Arguments

scores
A data.frame. Rownames have to be gene identifiers, followed by any number of columns. The column used for scoring is chosen by scoreColumn. See http://erminej.msl.ubc.ca/help/input-files/gene-scores/ for information abot how to specify scores. (for test = ORA, GSR and ROC)

scoreColumn
Integer or character. Which column of the scores data.frame to use as scores. Defaults to first column of scores. See http://erminej.msl.ubc.ca/help/input-files/gene-scores/ for details. (for test = ORA, GSR and ROC)

bigIsBetter
Logical. If TRUE large scores are considered to be higher. FALSE by default (as in p values).

logTrans
Logical. Should the data be -log10 transformed. Recommended for p values. FALSE by default

annotation
Annotation. A file path, a data.frame or a platform short name (eg. GPL127). If given a platform short name it will be downloaded from annotation repository of Pavlidis Lab (www.chibi.ubc.ca/microannots/). Note that if there is a file or folder with the same name as the platform name in the directory, that file will be read instead of getting a copy from Pavlidis Lab. If this file isn't a valid annotation file, the function will fail. If providing a custom annotation file, see erminej.msl.ubc.ca/help/input-files/gene-annotations/

aspects
Character vector. Which Go aspects to include in the analysis. Can be in long form (eg. 'Molecular Function') or short form (eg. c('M','C','B'))

geneReplicates
What to do when genes have multiple scores in input file (due to multiple probes per gene)

pAdjust
Which multiple test correction method to use. Can be "FDR" or 'Westfall-Young' (slower).

geneSetDescription
"Latest_GO", a file path that leads to a GO XML file or a URL that leads to a go ontology file that ends with rdf-xml.gz. Note that this is a mandatory argument. It defaults to Latest_GO but that option won't work if you don't have a working internet connection. See http://erminej.msl.ubc.ca/help/input-files/gene-set-descriptions/ for details

customGeneSets
Directory path or a named list of character strings. Use this option to create your own gene sets. If you provide directory you can specify probes or gene symbols to include in your gene sets. See http://erminej.msl.ubc.ca/help/input-files/gene-sets/ for information about format for this file. If you are providing a list, only gene symbols are accepted.

minClassSize
minimum class size

maxClassSize
maximum class size

output
Output file name.

return
If results should be returned. Set to FALSE if you only want a file

Details

The ROC is a well-known method for evaluating rankings of items, in this case genes. The ranking in this case comes from the gene scores. A gene set will get a good ROC if many genes in the gene set are near the top of the list.

The score measured for each gene set is the area under the ROC curve, a value between 0 and 1. If the genes in the gene set are randomly distributed in the ranking, you would expect a value near 0.5. Values near 1 indicate the genes in the gene set are near the top of the list, while values near 0 indicate the genes in the gene set are near the bottom of the list. In principle both values near 0 and near 1 are statistically significant, but p-values reported by ermineJ are based on the assumption that only the top of the list is of interest (e.g., we’re not considering “under-representation analysis”).

Unlike the other methods in ermineJ other than the PRC method, the ROC uses only the ranks of the gene scores. That is, all it cares about is the ordering of items obtained by your gene scores (e.g., t-test or fold-change), but doesn’t use the information about the relative values of the scores.

P-values for this analysis are computed using algorithms described in Breslin et al., 2004*. For more information on the ROC, you could do worse than reading the Wikipedia page http://en.wikipedia.org/wiki/Receiver_operator_characteristic.

Like other non-parametric techniques, using ranks costs some statistical power, but also makes fewer assumptions. Specifically, if you think the ordering of items in your data is more accurate than the actual p-values themselves, the ROC might be appropriate. The PRC method is similar in that it uses ranks, but puts more emphasis on genes in the set which are ranked very near the top. In contrast the ROC method looks at overall trends in the rankings.

Method overview taken from: http://erminej.msl.ubc.ca/help/tutorials/running-an-analysis-correlation/