scores

scores

scores(scores, scoreColumn = 1, bigIsBetter = FALSE, logTrans = FALSE)

Arguments

scores
A data.frame. Rownames have to be gene identifiers, followed by any number of columns. The column used for scoring is chosen by scoreColumn. See http://erminej.msl.ubc.ca/help/input-files/gene-scores/ for information abot how to specify scores. (for test = ORA, GSR and ROC)

scoreColumn
Integer or character. Which column of the scores data.frame to use as scores. Defaults to first column of scores. See http://erminej.msl.ubc.ca/help/input-files/gene-scores/ for details. (for test = ORA, GSR and ROC)

bigIsBetter
Logical. If TRUE large scores are considered to be higher. FALSE by default (as in p values).

logTrans
Logical. Should the data be -log10 transformed. Recommended for p values. FALSE by default